+++ lib/billy/capybara/rspec.rb
@@ -1,0 +1,5 @@
+require 'capybara/rspec'
+require 'billy/browsers/capybara'
+require 'billy/init/rspec'
+
+Billy::Browsers::Capybara.register_drivers