+++ lib/billy/handlers/cache_handler.rb
@@ -30,6 +30,10 @@ module Billy
             Billy.config.after_cache_handles_request.call(request, response)
           end
 
+          if Billy.config.cache_simulates_network_delays
+            Kernel.sleep(Billy.config.cache_simulates_network_delay_time)
+          end
+
           return response
         end
       end
@@ -46,8 +50,9 @@ module Billy
       request_uri = Addressable::URI.parse(url)
       if request_uri.query
         params = CGI.parse(request_uri.query)
-        if params['callback'].any? && response[:content].match(/\w+\(/)
-          response[:content].sub!(/\w+\(/, params['callback'].first + '(')
+        callback_name = Billy.config.dynamic_jsonp_callback_name
+        if params[callback_name].any? && response[:content].match(/\w+\(/)
+          response[:content].sub!(/\w+\(/, params[callback_name].first + '(')
         end
       end
     end