+++ lib/billy/init/rspec.rb
@@ -1,0 +1,15 @@
+module Billy
+  module RspecHelper
+    def proxy
+      Billy.proxy
+    end
+  end
+end
+
+RSpec.configure do |config|
+  config.include(Billy::RspecHelper)
+
+  config.after(:each) do
+    proxy.reset
+  end
+end